Move Classes/CheckOff* files to Classes/CheckOff/*
